# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScenarioCreateRequest": {
        "description": "Manual creation requires name+persona; `fromDraftId` short-circuits\nthat (B1-01: idempotent via scenarios.source_draft_ref UNIQUE).",
        "properties": {
          "assertions": {
            "items": {},
            "title": "Assertions",
            "type": "array"
          },
          "fromDraftId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fromdraftid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "personaInitials": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personainitials"
          },
          "script": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script"
          }
        },
        "title": "ScenarioCreateRequest",
        "type": "object"
      },
      "ScenarioSummary": {
        "properties": {
          "assertCount": {
            "title": "Assertcount",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "persona": {
            "title": "Persona",
            "type": "string"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "string"
          },
          "status": {
            "enum": [
              "pass",
              "warn",
              "fail",
              "idle"
            ],
            "title": "Status",
            "type": "string"
          },
          "suiteId": {
            "title": "Suiteid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "suiteId",
          "name",
          "persona",
          "assertCount",
          "status",
          "score",
          "runId"
        ],
        "title": "ScenarioSummary",
        "type": "object"
      },
      "ScenarioUpdate": {
        "properties": {
          "assertions": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assertions"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "script": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script"
          }
        },
        "title": "ScenarioUpdate",
        "type": "object"
      },
...
      "SuiteCreate": {
        "properties": {
          "agentId": {
            "title": "Agentid",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "agentId"
        ],
        "title": "SuiteCreate",
        "type": "object"
      },
      "SuiteDetail": {
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "desc": {
            "title": "Desc",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastRun": {
            "title": "Lastrun",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "passRate": {
            "title": "Passrate",
            "type": "string"
          },
          "pr": {
            "title": "Pr",
            "type": "integer"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioSummary"
            },
            "title": "Scenarios",
            "type": "array"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "desc",
          "agent",
          "lastRun",
          "passRate",
          "pr",
          "count",
          "scenarios"
        ],
        "title": "SuiteDetail",
        "type": "object"
      },
      "SuiteListItem": {
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "desc": {
            "title": "Desc",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastRun": {
            "title": "Lastrun",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "passRate": {
            "title": "Passrate",
            "type": "string"
          },
          "pr": {
            "title": "Pr",
            "type": "integer"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "name",
          "desc",
          "agent",
          "lastRun",
          "passRate",
          "pr",
          "count"
        ],
        "title": "SuiteListItem",
        "type": "object"
      },
      "SuiteUpdate": {
        "properties": {
          "agentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agentid"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "SuiteUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "cadence-brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/scenarios/{scenario_id}": {
      "delete": {
        "operationId": "delete_scenario",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scenario",
        "tags": [
          "suites"
        ]
      },
      "patch": {
        "operationId": "update_scenario",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scenario",
        "tags": [
          "suites"
        ]
      }
    },
...
    "/v1/suites": {
      "get": {
        "operationId": "list_suites",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SuiteListItem"
                  },
                  "title": "Response List Suites",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Suites",
        "tags": [
          "suites"
        ]
      },
      "post": {
        "operationId": "create_suite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuiteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteListItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Suite",
        "tags": [
          "suites"
        ]
      }
    },
    "/v1/suites/{suite_id}": {
      "delete": {
        "operationId": "delete_suite",
        "parameters": [
          {
            "in": "path",
            "name": "suite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Suite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Suite",
        "tags": [
          "suites"
        ]
      },
      "get": {
        "operationId": "get_suite",
        "parameters": [
          {
            "in": "path",
            "name": "suite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Suite Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Suite",
        "tags": [
          "suites"
        ]
      },
      "patch": {
        "operationId": "update_suite",
        "parameters": [
          {
            "in": "path",
            "name": "suite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Suite Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuiteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Suite",
        "tags": [
          "suites"
        ]
      }
    },
    "/v1/suites/{suite_id}/scenarios": {
      "post": {
        "operationId": "add_scenario",
        "parameters": [
          {
            "in": "path",
            "name": "suite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Suite Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Scenario",
        "tags": [
          "suites"
        ]
      }
    }
  }
}